requirements.txt

streamlit
pandas
openpyxl


numpy
scikit-learn
sentence-transformers
openai
python-dotenv


torch @ https://download.pytorch.org/whl/cpu/torch-2.3.1%2Bcpu-cp311-cp311-linux_x86_64.whl
